Agent tester added. Class structure changed.
